requirements.txt
bcrypt
pycryptodome
customtkinter
pycryptodome